Changed the rmdir line to fix an error thrown when the out folder doesn't exist. Closes #7

+++ demos/pwm-control/hardware-test.bat
@@ -1,5 +1,3 @@
-@echo off
-
 SET scriptbuildflag=--release
 SET scriptbin=pwm_control
 IF "%1"=="--debug" (SET "scriptbuildflag=")
@@ -9,7 +7,7 @@ IF NOT DEFINED scriptbuildflag (
 	SET scriptbuildmode=release
 )
 cargo build %scriptbuildflag% || goto :error
-rmdir /S /Q out || goto :error
+if exist out rmdir /S /Q out || goto :error
 mkdir out || goto :error
 copy "target\thumbv7em-none-eabihf\%scriptbuildmode%\%scriptbin%" "out\%scriptbin%" || goto :error
 REM arm-none-eabi-objdump -d -S -C "out\%scriptbin%" > "out\%scriptbin%.lst" || goto :error
